Merge in nodelibs

+++ nix-libs/nodeLib/installBinaries.py
@@ -3,31 +3,32 @@
 # See: https://docs.npmjs.com/files/package.json#bin
 import json
 import os
-from os.path import join, isdir, normpath
+from os.path import exists, join, isdir, normpath
+import stat
 import sys
 
-with open('package.json', 'r') as f:
+with open("package.json", "r") as f:
     package = json.load(f)
 
-if 'bin' not in package:
+if "bin" not in package:
     sys.exit(0)
 
-_bin = package['bin']
+_bin = package["bin"]
 
 if isinstance(_bin, basestring):
     # This is equivalent to a singleton dictionary where the key is the
     # name of the package.
-    _bin = {package['name']: _bin}
+    _bin = {package["name"]: _bin}
 elif not isinstance(_bin, dict):
     # Otherwise it must be a dictionary.
     sys.exit("Expected `bin` key in package.json to point to a string "
              "or a dict, but it is '{}', of type '{}'"
              .format(_bin, type(_bin).__name__))
 
-out_dir = os.environ['out']
+out_dir = os.environ["out"]
 
 # Create the .bin folder
-bin_folder = join(out_dir, 'bin')
+bin_folder = join(out_dir, "bin")
 if not isdir(bin_folder):
     os.makedirs(bin_folder)
 
@@ -35,7 +36,14 @@ print("Creating binaries in {}".format(bin_folder))
 
 for bin_name, bin_path in _bin.items():
     # Get the absolute path of the script being pointed to.
-    bin_abs_path = normpath(join(out_dir, 'lib', 'node_modules',
-                            package['name'], bin_path))
+    bin_abs_path = normpath(join(out_dir, "lib", "node_modules",
+                            package["name"], bin_path))
+    if not exists(bin_abs_path):
+        sys.exit("Package {} version {} declares a binary {} at path {}, "
+                 "but there is no such file at that path.".format(
+                     package["name"], package["version"], bin_name, bin_path))
     print("Linking binary {} to {}".format(bin_name, bin_abs_path))
+    # Add executable permissions to the binary.
+    bin_stats = os.stat(bin_abs_path)
+    os.chmod(bin_abs_path, bin_stats.st_mode | stat.S_IEXEC)
     os.symlink(bin_abs_path, join(bin_folder, bin_name))
